requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg[binary]>=3.1.0
alembic==1.12.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
authlib==1.2.1
python-owasp-zap-v2.4==0.0.2

# Validation
email-validator==2.1.0
pydantic==2.13.2
pydantic-settings==2.1.0

# Email
aiosmtplib==2.0.1
jinja2==3.1.2

# Rate Limiting & Cache
redis==5.0.1
slowapi==0.1.9

# Task Queue
celery==5.3.4

# Security & Sanitization
bleach==6.1.0
itsdangerous==2.1.2

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
